be/requirements.txt
# Core Flask Dependencies
Flask==3.0.2
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Werkzeug==3.0.1

# Environment Management
python-dotenv==1.0.1

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.11.0
PyMuPDF==1.23.26
reportlab==4.1.0

# AI/ML - Ollama Integration
ollama==0.1.6
requests==2.31.0

# Data Validation
pydantic==2.6.1
marshmallow==3.20.2

# Security & Rate Limiting
Flask-Limiter==3.5.0
itsdangerous==2.1.2

# Caching
Flask-Caching==2.1.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Code Quality
black==24.2.0
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0

# Pre-commit Hooks
pre-commit==3.6.0

# Production Server
gunicorn==21.2.0

# Utilities
python-magic-bin==0.4.14  # For Windows file type detection
